only pass lines between each pass function
In order to better decouple the passes, only pass lines between each
pass function. This will allow for better testing and easier
debugging.

This does mean that each pass function will need to be responsible for
re-parsing the lines into the appropriate data structures that each
stores internally.

Remove the references and targets from argument lists for the pass functions.